Refactor coordinates search
