+++ src/ui/components/SearchCoordinates/template.hbs
@@ -1,0 +1,11 @@
+<div>
+  <button
+    class="use-location"
+    disabled={{@isDisabled}}
+    type="button"
+    onclick={{action searchByLocation}}
+    data-test-search-near
+  >
+    Show nearby locations
+  </button>
+</div>
