+++ src/ui/components/SearchForm/template.hbs
@@ -5,5 +5,4 @@
       Find location
     </button>
   </form>
-  <button class="use-location" disabled={{@isDisabled}} type="button" onclick={{action searchByLocation}} data-test-search-near>Show nearby locations</button>
 </div>
